+++ Changelog.md
@@ -1,3 +1,12 @@
+## Version 1.1.10
+### Fixes
+- Filtered out devices without serialport that caused crash #49
+
+## Version 1.1.9
+### Updates
+- Updated CEREG notification level to 5 #47
+- Updated auto device filter for macOS #47
+
 ## Version 1.1.8
 ### Updates
 - Updated according to changes of Electron dialog API